Blog Listing Page Now Dynamic
Made the blog listing page dynamic, meaning that and the single page blog blades are both using dynamic data instead of hardcoded ones. Yay!

Still some problems with dating and metadata. Heading into using YamlFrontMatter and using that for file metadata.

This is through Episode 11 of the Laravel tutorial on Laracasts.