Added Metadata to posts, and learned YamlFrontMatter and Collections in Laravel
We've done some work with YamlFrontMatter to add metadata to all of our posts, and modified our blade template to output those posts on the homepage.

Also, we're adding further variables and augmenting the all() method in the Post() class.

Now we're moving from array_map to using collections in Laravel, which gives you various object oriented methods to interact with arrays.

This is most of the way through episode 12 of the Laracasts Laravel tutorial. Just need to update the single post blade to use the new Laravel features we worked with in order to finish this lesson.