Finished EP 15 and 16 of Tutorial; layout files, components, and small tweaks
Broke out our main layout into both a component file and a template layout file, just so we know how to do both. In this application, the author of the tutorial said either could be the "right" way to do this sorta thing. But I think that in the future we'll see each method have it's own set of advantages/disadvantages. Or maybe it's just syntax.

Using components reminds me of Angular syntax, though, so maybe I'll go for those in the future if they remain equally important.

Also added some final tweaks; removed the "where" clause in the routes file in favor of throwing an exception within the find() method.

In the Laracasts tutorial, the author said that in certain cases you might want to have one method find() and one method findOrFail(), the latter one throwing an exception while the first one just fails by returning null. This doesn't really make sense to me in this application, as why would you want an ugly error to ever be displayed? Maybe this will become more relevant as we continue in the tutorial.

It's also possible that you might want this in Models that have a more backend component, however. Maybe in certain cases, you want this to send "null" so you can check for that in your own methods instead of the thing failing automatically.

For now, I've left those revised methods commented out in case we need them later. But, while we're just modifying this as we are right now, it's not entirely necessary.